Subscribe callbacks, not components; return and call a `remove` method on each subscription
